Added a few global variables
requestData holds all data that the "black box" needs, and more.

currentSortInfo holds the current information for sorting the lists we create. It will remember the list sortings even after we maybe contact the database for other things exept sorting or building lists.

outCart is more or less a shopping cart, it keeps all information about the things in the cart.